update replace file
