+++ handlers/assessment/assessment_article.go
@@ -1,4 +1,4 @@
-package assessment_article
+package assessment
 
 import (
 	"fmt"
