+++ main.go
@@ -1,11 +1,8 @@
 package main
 
 import (
+	"CRUD-API/api/middlewares"
 	"CRUD-API/handlers/assessment"
-	"CRUD-API/handlers/assessment_article"
-	"CRUD-API/handlers/assessment_progress"
-	"CRUD-API/handlers/assessment_project"
-	"CRUD-API/handlers/assessment_report"
 	"CRUD-API/handlers/auth"
 	"CRUD-API/handlers/degree"
 	"CRUD-API/handlers/experience"
@@ -19,7 +16,6 @@ import (
 	"CRUD-API/handlers/researcher_list"
 	"CRUD-API/handlers/user"
 	"CRUD-API/initializers"
-	"CRUD-API/middlewares"
 
 	// . "CRUD-API/models"
 
@@ -130,7 +126,7 @@ func main() {
 	r.DELETE("/assessment/:id", assessmentHandler.DeleteAssessmentHandler)
 
 	//Article Zones
-	articleHandler := assessment_article.NewArticleHandler(db)
+	articleHandler := assessment.NewArticleHandler(db)
 	r.GET("/article/", articleHandler.ListArticle)
 	r.GET("/article/:id", articleHandler.GetArticleHandler)
 	r.POST("/article", articleHandler.CreateArticleHandler)
@@ -138,7 +134,7 @@ func main() {
 	r.DELETE("/article/:id", articleHandler.DeleteArticleHandler)
 
 	//AssessmentProgress Zones
-	progressHandler := assessment_progress.NewProgressHandler(db)
+	progressHandler := assessment.NewProgressHandler(db)
 	r.GET("progress/", progressHandler.ListProgress)
 	r.GET("/progress/:id", progressHandler.GetProgressHandler)
 	r.POST("/progress", progressHandler.CreateProgressHandler)
@@ -146,7 +142,7 @@ func main() {
 	r.DELETE("/progress/:id", progressHandler.DeleteProgressHandler)
 
 	//AssessmentReport Zones
-	reportHandler := assessment_report.NewReportHandler(db)
+	reportHandler := assessment.NewReportHandler(db)
 	r.GET("report/", reportHandler.ListReport)
 	r.GET("/report/:id", reportHandler.GetReportHandler)
 	r.POST("/report", reportHandler.CreateReportHandler)
@@ -154,7 +150,7 @@ func main() {
 	r.DELETE("/report/:id", reportHandler.DeleteReportHandler)
 
 	//Assessment Project Zones
-	projectHandler := assessment_project.NewProjectHandler(db)
+	projectHandler := assessment.NewProjectHandler(db)
 	r.GET("/project", projectHandler.ListProjects)
 	r.GET("/project/:id", projectHandler.GetProjectHandler)
 	r.POST("/project", projectHandler.CreateProjectHandler)
